+++ server/http/get.js
@@ -1,0 +1,4 @@
+﻿var request = require("./request");
+module.exports = function (type, url, options) {
+  return request("GET", type || "text/plain", url, options);
+};