+++ src/fx/HTTP/getBuffer.js
@@ -1,12 +1,10 @@
 ﻿pliny.function("Primrose.HTTP", {
   name: "getBuffer",
   description: "Get an ArrayBuffer from a server.",
-  returns: "If no success function is provide, returns a Promise. Otherwise, returns undefined.",
+  returns: "Promise",
   parameters: [
     { name: "url", type: "String", description: "The resource to which the request is being sent." },
-    { name: "progress", type: "Function", description: "(Optional) A callback function to be called as the download from the server progresses." },
-    { name: "success", type: "Function", description: "(Optional) the callback to issue whenever the request finishes successfully, even going so far as to check HTTP status code on the OnLoad event." },
-    { name: "error", type: "Function", description: "(Optional) the callback to issue whenever an error occurs." }
+    { name: "options.progress", type: "Function", description: "(Optional) A callback function to be called as the download from the server progresses." }
   ],
   examples: [{
     name: "Make a GET request for an ArrayBuffer.",
@@ -27,6 +25,6 @@
       console.error.bind(console, \"error loading\")\n"}
   ]
 });
-Primrose.HTTP.getBuffer = function (url, progress, success, error) {
-  return Primrose.HTTP.get("arraybuffer", url, progress, success, error);
+Primrose.HTTP.getBuffer = function (url, options) {
+  return Primrose.HTTP.get("arraybuffer", url, options);
 };